Intersect line segment with rectangle
Check whether any side (segment) of the rectangle intersects with the given
segment. Segments intersect if the two points of the other segment are on
each side when looking from first point of rectangle segment to the other.
This is the case when one is counterclockwise from rectangle segment and the
other is clockwise. This can be determined by calculating the slope from one
point to the next and from the next to the last. This segment intersection
test is performed for all sides of the rectangle.