Draw tree and add nodes every update call
drawTree function defined in rer_tree.h and rer_tree.cpp draws tree by drawing
a line from every parent to its children. Samples are added every update call
by taking a random point on the screen which is not in a rectangle and find
the nearest tree node. If direct_line flag ist set, there's a connection from
the nearest node to the new node on the sample pos, otherwise a new node is
added to the tree in an incremental distance from the nearest node in the
direction of the sample.