Connect to end directly if possible and color final path
Every update call before sampling a new random position a test is made if
it is possible to connect the nearest node to the end directly. If this is
the case the simulation is stopped and the final path is colored red by
following the end_node's parent and setting the color until the root node
is reached.